Small fixes to make the code safer
JsonTree::root renamed to JsonTree:root_ to not to shadow Node root variablie, which is passed to the ctor as a parameter
Now we use fuzzyCompare instead of == to compare floats in debugging operators ==
Now we store road distance in size_t
"u" prefix was add to avoid comparing between signed and unsigned integer in tests 
unused operators were removed
useless ';' were removed

+++ src/routeDistancesDict.h
@@ -25,4 +25,4 @@ struct FromToHasher
     }
 };
 
-using RouteDistancesMap = std::unordered_map<FromTo, unsigned int, FromToHasher>;
+using RouteDistancesMap = std::unordered_map<FromTo, size_t, FromToHasher>;
